Mémos - Fix for generated inputs binded with iCheck

+++ resources/views/memos/list.blade.php
@@ -1,14 +1,18 @@
-@foreach($memos as $memo)
-  <?php $check = ($memo->active == 0)? 0 : 1; ?>
-<li class="list-group-item" data-id="{{ $memo->id }}"><i class=" glyphicon glyphicon-th"></i>
-  {!! Form::open(['route' => ['memo.check', $memo->id], 'method' => 'POST', 'class' => 'check-element']) !!}
-  {!! Form::checkbox('check', 1, $check) !!}
-  {!! Form::close() !!}
-  <span class="edit-element">{{ $memo->name }}</span>
-    {!! Form::open(['route' => ['project.memo.destroy', $project->slug, $memo->id], 'method' => 'DELETE', 'class' => 'delete-element pull-right']) !!}
-		<div class="form-group">
-		    {!! Form::submit('×', ['class' => 'close'])!!}
-		</div>
-    {!!  Form::close() !!}
-</li>
-@endforeach
+@if(count($memos))
+  @foreach($memos as $memo)
+    <?php $check = ($memo->active == 0)? 0 : 1; ?>
+  <li class="list-group-item" data-id="{{ $memo->id }}"><i class=" glyphicon glyphicon-th"></i>
+    {!! Form::open(['route' => ['memo.check', $memo->id], 'method' => 'POST', 'class' => 'check-element']) !!}
+    {!! Form::checkbox('check', 1, $check) !!}
+    {!! Form::close() !!}
+    <span class="edit-element">{{ $memo->name }}</span>
+      {!! Form::open(['route' => ['project.memo.destroy', $project->slug, $memo->id], 'method' => 'DELETE', 'class' => 'delete-element pull-right']) !!}
+  		<div class="form-group">
+  		    {!! Form::submit('×', ['class' => 'close'])!!}
+  		</div>
+      {!!  Form::close() !!}
+  </li>
+  @endforeach
+@else
+  <li class="no-result">{!! trans('memo.no_memo_list') !!}</li>
+@endif
